+++ CHANGELOG.md
@@ -7,6 +7,14 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
+- Add `__xstat` hook, fixes #867
+
+### Fixed
+
+- getaddrinfo: if node is NULL just bypass, as it's just for choosing ip/port, Fixes#858 and #848
+
 ## 3.14.3
 
 ### Fixed
@@ -18,7 +26,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Fixed
 
 - vscode ext: use process env for running mirrord. Fixes [#854](https://github.com/metalbear-co/mirrord/issues/854)
-- getaddrinfo: if node is NULL just bypass, as it's just for choosing ip/port, Fixes[#858](https://github.com/metalbear-co/mirrord/issues/858) and [#848](https://github.com/metalbear-co/mirrord/issues/848)
 
 ## 3.14.1